fix(agent): a uint32 wire length is -1 on 32-bit, and the suite never ran there (D362)
Fuzzing the privileged parse surface. The proposal predicted zero defects and said
so up front; it was wrong, and the prediction is left in the record because being
wrong in that direction is the result.

`int(someUint32FromTheWire)` is -1 on a 32-bit platform. Every "is this length past
the ceiling?" check then passes and the following slice panics:

  execmon.decodeMeta      buf[4294967295:]                 FuzzDecodeMeta on 386
  openmon.decodeMeta      same, in the FAN_OPEN_PERM loop   FuzzDecodeMeta on 386
  openipc.ReadRequest     short alloc, then body[:pathLen]  an EXISTING test on 386
  the inotify record walk buf[16:15]                        reading it while extracting it

THE REAL FINDING IS NOT THREE BUGS. It is that the suite had never run on a 32-bit
architecture, while GOARCH=386 and GOARCH=arm both compile the agent today. The
openipc site was already covered by a test named
TestADeclaredLengthBeyondTheBoundIsRefusedBeforeAllocating — the exact property that
fails — and it passed on amd64 forever. So the durable fix is a CI step, not four
patches.

Fuzzing alone would not have found all of it: two came from the fuzzer, one from an
existing test under 386, one from reading the arithmetic. On amd64 every one of them
is correct.

The roadmap's justification for this ticket is also corrected. "ClamAV-CVE-class RCE"
is not the threat — Go is memory-safe. What these decoders can do is panic, allocate
unboundedly, or fail to advance, and the severity comes from the process: the agent
answers BLOCKING permission events, so a crash or a spin leaves openers in
uninterruptible windows with the gate failing open.

Seven targets assert PROGRESS and declared bounds, not merely absence of panic. All
mutation-verified — including one assertion that could NOT fail until a seed carrying
an over-ceiling body was added, which the mutation is what exposed. ~380M executions
after the fixes, no crashers, so testdata/fuzz/ is empty; the CI step is a labelled
smoke test, and the durable value is that go test replays seeds forever.

SEPARATELY, AND BIGGER: the worker REFUSES TO START on linux/386 — its seccomp
denylist names `accept`, not a syscall there, so the filter will not assemble. And it
would be ineffective anyway, because i386 socket ops go through `socketcall`, which
the list does not name. The assembly failure is what holds the boundary, by refusing
to parse. Right direction, reached by accident. NOT fixed here: a sandbox that looks
applied and is not is worse than a platform that refuses to run. Recorded as an owner
decision on the roadmap.

+++ internal/agent/execmon/execmon.go
@@ -52,10 +52,21 @@ func decodeMeta(buf []byte) (m meta, rest []byte, ok bool) {
 	m.Mask = binary.LittleEndian.Uint64(buf[8:16])
 	m.FD = int32(binary.LittleEndian.Uint32(buf[16:20]))
 	m.PID = int32(binary.LittleEndian.Uint32(buf[20:24]))
-	if m.EventLen < metaLen || int(m.EventLen) > len(buf) {
+	// COMPARED IN uint64, NOT int, AND THIS IS A FIX. `int(m.EventLen) > len(buf)` was the original, and
+	// on a 32-bit platform `int` is 32 bits, so an event_len of 0xFFFFFFFF converts to -1: the
+	// over-run check reads `-1 > 24` and passes, the under-run check reads `0xFFFFFFFF < 24` on the
+	// UNSIGNED field and also passes, and the slice below panics with `[4294967295:24]`. The agent
+	// builds for linux/386 and linux/arm today. Found by FuzzDecodeMeta under GOARCH=386; on amd64 the
+	// same input is rejected correctly, which is why nothing had noticed.
+	//
+	// The order of the conversion is what decides this. `connectors/fanotify.ParseEvent` converts to int
+	// BEFORE comparing, so its negative value trips the under-run check and it is accidentally safe —
+	// same class, opposite outcome, and neither site says which discipline it is relying on.
+	if m.EventLen < metaLen || uint64(m.EventLen) > uint64(len(buf)) {
 		return m, nil, false // a length that under/over-runs the buffer is malformed
 	}
-	return m, buf[m.EventLen:], true
+	// Safe as an int now: the guard above proved EventLen fits within len(buf).
+	return m, buf[int(m.EventLen):], true
 }
 
 // DenyEvaluator is the pure, parser-free inline exec decider (satisfies watchdog.Evaluator).
